PMM-7590 Refactoring
