+++ services/management/jobs.go
@@ -24,17 +24,17 @@ import (
 	"gopkg.in/reform.v1"
 
 	"github.com/percona/pmm-managed/models"
-	"github.com/percona/pmm-managed/services/jobs"
+	"github.com/percona/pmm-managed/services/agents"
 )
 
 // JobsAPIService provides methods for Jobs starting and management.
 type JobsAPIService struct {
 	db          *reform.DB
-	jobsService *jobs.Service
+	jobsService *agents.JobsService
 }
 
 // NewJobsAPIServer creates new jobs service.
-func NewJobsAPIServer(db *reform.DB, service *jobs.Service) *JobsAPIService {
+func NewJobsAPIServer(db *reform.DB, service *agents.JobsService) *JobsAPIService {
 	return &JobsAPIService{
 		db:          db,
 		jobsService: service,
